sync-engine - Reorganize
