+++ modules/util/sync-engine/src/com/liferay/sync/engine/filesystem/listener/SyncSiteWatchEventListener.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.sync.engine.filesystem;
+package com.liferay.sync.engine.filesystem.listener;
 
 import com.liferay.sync.engine.model.SyncAccount;
 import com.liferay.sync.engine.model.SyncFile;
